mm_sim: Use operator info provided by SimManager
Sim.OperatorIdentifier and Sim.PreferredNetworks should export values provided by org.ofono.SimManager as the conditions of the current network may vary (eg. when roaming).

Additionally, do not provide Sim.OperatorName as oFono does not provide corresponding information. The previous method used does not work appropriately due to the current network being not always being the home network.

As previously, change the OperatorIdentifier to be in format MCCMNC as it should be.